fix dialog not mounting
